automatically backed up by learn

+++ lib/bachelor.rb
@@ -53,7 +53,7 @@ def get_average_age_for_season(data, season)
   data.each do |season_num, season_attrib|
     if season_num == season
       season_attrib.each do |obj|
-        total_age += obj["age"].to_i
+        total_age += obj["age"].to_f
         total_contestants += 1
       end
     end
